+++ app/models/formats/ogr.js
@@ -9,16 +9,6 @@ var spawn       = require('child_process').spawn
 // Keeps track of what's waiting baking for export
 var bakingExports = {};
 
-// Return database username from user_id
-// NOTE: a "null" user_id is a request to use the public user
-function userid_to_dbuser(user_id) {
-  if ( _.isString(user_id) )
-      return _.template(global.settings.db_user, {user_id: user_id});
-  return "publicuser" // FIXME: make configurable
-};
-
-
-
 function ogr(id) {
   this.id = id;
 }
@@ -43,9 +33,10 @@ ogr.prototype = {
 
   getKey: function(options) {
     return [this.id,
-        options.dbname,
-        options.user_id,
+        options.dbopts.dbname,
+        options.dbopts.user,
         options.gn,
+        this.generateMD5(options.filename),
         this.generateMD5(options.sql)].concat(options.skipfields).join(':');
   },
 
@@ -58,14 +49,27 @@ ogr.prototype = {
 };
 
 // Internal function usable by all OGR-driven outputs
-ogr.prototype.toOGR = function(dbname, user_id, gcol, sql, skipfields, out_format, out_filename, callback) {
+ogr.prototype.toOGR = function(options, out_format, out_filename, callback) {
+
+  var gcol = options.gn;
+  var sql = options.sql;
+  var skipfields = options.skipfields;
+  var out_layername = options.filename;
+
+  var dbopts = options.dbopts;
+
   var ogr2ogr = 'ogr2ogr'; // FIXME: make configurable
-  var dbhost = global.settings.db_host;
-  var dbport = global.settings.db_port;
-  var dbuser = userid_to_dbuser(user_id);
-  var dbpass = ''; // turn into a parameter..
+  var dbhost = dbopts.host;
+  var dbport = dbopts.port; 
+  var dbuser = dbopts.user; 
+  var dbpass = dbopts.pass;
+  var dbname = dbopts.dbname; 
+
+  var that = this;
 
   var columns = [];
+  var geocol;
+  var pg;
 
   // Drop ending semicolon (ogr doens't like it)
   sql = sql.replace(/;\s*$/, ''); 
@@ -73,32 +77,62 @@ ogr.prototype.toOGR = function(dbname, user_id, gcol, sql, skipfields, out_forma
   Step (
 
     function fetchColumns() {
-      var colsql = 'SELECT * FROM (' + sql + ') as _cartodbsqlapi LIMIT 1';
-      var pg = new PSQL(user_id, dbname, 1, 0);
+      var colsql = 'SELECT * FROM (' + sql + ') as _cartodbsqlapi LIMIT 0';
+      pg = new PSQL(dbopts);
       pg.query(colsql, this);
     },
-    function spawnDumper(err, result) {
+    function findSRS(err, result) {
       if (err) throw err;
 
       //if ( ! result.rows.length ) throw new Error("Query returns no rows");
 
-      // Skip system columns
-      if ( result.rows.length ) {
-        for (var k in result.rows[0]) {
-          if ( skipfields.indexOf(k) != -1 ) continue;
-          if ( out_format != 'CSV' && k == "the_geom_webmercator" ) continue; // TODO: drop ?
-          if ( out_format == 'CSV' ) columns.push('"' + k + '"::text');
-          else columns.push('"' + k + '"');
+      var needSRS = that._needSRS;
+
+      // Skip system columns, find geom column
+      for (var i=0; i<result.fields.length; ++i) {
+        var field = result.fields[i];
+        var k = field.name;
+        if ( skipfields.indexOf(k) != -1 ) continue;
+        if ( out_format != 'CSV' && k == "the_geom_webmercator" ) continue; // TODO: drop ?
+        if ( out_format == 'CSV' ) columns.push(pg.quoteIdentifier(k)+'::text');
+        else columns.push(pg.quoteIdentifier(k));
+
+        if ( needSRS ) {
+          if ( ! geocol && pg.typeName(field.dataTypeID) == 'geometry' ) {
+            geocol = k
+          }
         }
-      } else columns.push('*');
+      }
       //console.log(columns.join(','));
+
+      if ( ! needSRS || ! geocol ) return null;
 
       var next = this;
 
-      sql = 'SELECT ' + columns.join(',')
+      var qgeocol = pg.quoteIdentifier(geocol);
+      var sridsql = 'SELECT ST_Srid(' + qgeocol + ') as srid, GeometryType(' +
+                   qgeocol + ') as type FROM (' + sql + ') as _cartodbsqlapi WHERE ' +
+                   qgeocol + ' is not null limit 1';
+
+      pg.query(sridsql, function(err, result) {
+        if ( err ) { next(err); return; }
+        if ( result.rows.length ) {
+          var srid = result.rows[0].srid;
+          var type = result.rows[0].type;
+          next(null, srid, type);
+        }
+      });
+
+    },
+    function spawnDumper(err, srid, type) {
+      if (err) throw err;
+
+      var next = this;
+
+      var ogrsql = 'SELECT ' + columns.join(',')
           + ' FROM (' + sql + ') as _cartodbsqlapi';
 
-      var child = spawn(ogr2ogr, [
+      var ogrargs = [
         '-f', out_format,
         '-lco', 'ENCODING=UTF-8',
         '-lco', 'LINEFORMAT=CRLF',
@@ -107,22 +141,27 @@ ogr.prototype.toOGR = function(dbname, user_id, gcol, sql, skipfields, out_forma
          + " user=" + dbuser
          + " dbname=" + dbname
          + " password=" + dbpass
-         + " tables=fake" // trick to skip query to geometry_columns
+         + " tables=fake" // trick to skip query to geometry_columns (private)
+                          // in turn breaks knowing SRID with gdal-0.10.1:
+                          // http://github.com/CartoDB/CartoDB-SQL-API/issues/110
          + "",
-        '-sql', sql
-      ]);
+        '-sql', ogrsql
+      ];
+
+      if ( srid ) {
+        ogrargs.push('-a_srs', 'EPSG:'+srid);
+      }
+
+      if ( type ) {
+        ogrargs.push('-nlt', type);
+      }
+
+      ogrargs.push('-nln', out_layername);
+
+      var child = spawn(ogr2ogr, ogrargs);
 
 /*
-console.log(['ogr2ogr',
-        '-f', '"'+out_format+'"',
-        out_filename,
-        "'PG:host=" + dbhost
-         + " user=" + dbuser
-         + " dbname=" + dbname
-         + " password=" + dbpass
-         + " tables=fake" // trick to skip query to geometry_columns
-         + "'",
-        "-sql '", sql, "'"].join(' '));
+console.log('ogr2ogr ' + _.map(ogrargs, function(x) { return "'" + x + "'"; }).join(' '));
 */
 
       var stdout = '';
@@ -142,7 +181,8 @@ console.log(['ogr2ogr',
 
       child.on('exit', function(code) {
         if ( code ) {
-          var emsg = stderr.split('\n')[0];
+          console.log("ogr2ogr exited with code " + code);
+          var emsg = stderr ? stderr.split('\n')[0] : ( "unknown ogr2ogr error (code " + code + ")" );
           // TODO: add more info about this error ?
           //if ( RegExp(/attempt to write non-.*geometry.*to.*type shapefile/i).exec(emsg) )
           next(new Error(emsg));
@@ -157,15 +197,24 @@ console.log(['ogr2ogr',
   );
 };
 
-ogr.prototype.toOGR_SingleFile = function(dbname, user_id, gcol, sql, skipfields, fmt, ext, callback) {
+ogr.prototype.toOGR_SingleFile = function(options, fmt, callback) {
+
+  var dbname = options.dbopts.dbname;
+  var user_id = options.dbopts.user;
+  var gcol = options.gcol;
+  var sql = options.sql;
+  var skipfields = options.skipfields;
+  var ext = this._fileExtension;
+  var layername = options.filename;
+
   var tmpdir = global.settings.tmpDir || '/tmp';
-  var reqKey = [ fmt, dbname, user_id, gcol, this.generateMD5(sql) ].concat(skipfields).join(':');
+  var reqKey = [ fmt, dbname, user_id, gcol, this.generateMD5(layername), this.generateMD5(sql) ].concat(skipfields).join(':');
   var outdirpath = tmpdir + '/sqlapi-' + process.pid + '-' + reqKey;
   var dumpfile = outdirpath + ':cartodb-query.' + ext;
 
   // TODO: following tests:
   //  - fetch query with no "the_geom" column
-  this.toOGR(dbname, user_id, gcol, sql, skipfields, fmt, dumpfile, callback);
+  this.toOGR(options, fmt, dumpfile, callback);
 };
 
 ogr.prototype.sendResponse = function(opts, callback) {