+++ test/prepare_db.sh
@@ -11,16 +11,39 @@ TESTENV=../config/environments/test.js
 
 
 # Extract postgres configuration
-#PGUSER=`grep \.db_user ${TESTENV} | sed "s/.*= *'\([^']*\)'.*/\1/"`
-#echo "PGUSER: [$PGUSER]"
-PGHOST=`grep \.db_host ${TESTENV} | sed "s/.*= *'\([^']*\)'.*/\1/"`
+PGHOST=`node -e "console.log(require('${TESTENV}').db_host || '')"`
 echo "PGHOST: [$PGHOST]"
-PGPORT=`grep \.db_port ${TESTENV} | sed "s/.*=[\t ]*'\([^']*\)'.*/\1/"`
+PGPORT=`node -e "console.log(require('${TESTENV}').db_port || '')"`
 echo "PGPORT: [$PGPORT]"
 
+PUBLICUSER=`node -e "console.log(require('${TESTENV}').db_pubuser || 'xxx')"`
+PUBLICPASS=`node -e "console.log(require('${TESTENV}').db_pubuser_pass || 'xxx')"`
+echo "PUBLICUSER: [${PUBLICUSER}]"
+echo "PUBLICPASS: [${PUBLICPASS}]"
 
-TEST_DB="cartodb_test_user_1_db"
-REDIS_PORT=6333 # TODO: read from environment file
+
+TESTUSERID=1
+
+TESTUSER=`node -e "console.log(require('${TESTENV}').db_user || '')"`
+if test -z "$TESTUSER"; then
+  echo "Missing db_user from ${TESTENV}" >&2
+  exit 1
+fi
+TESTUSER=`echo ${TESTUSER} | sed "s/<%= user_id %>/${TESTUSERID}/"`
+echo "TESTUSER: [${TESTUSER}]"
+
+TESTPASS=`node -e "console.log(require('${TESTENV}').db_user_pass || '')"`
+TESTPASS=`echo ${TESTPASS} | sed "s/<%= user_id %>/${TESTUSERID}/"`
+echo "TESTPASS: [${TESTPASS}]"
+
+TEST_DB=`node -e "console.log(require('${TESTENV}').db_base_name || '')"`
+if test -z "$TEST_DB"; then
+  echo "Missing db_base_name from ${TESTENV}" >&2
+  exit 1
+fi
+TEST_DB=`echo ${TEST_DB} | sed "s/<%= user_id %>/${TESTUSERID}/"`
+
+REDIS_PORT=`node -e "console.log(require('${TESTENV}').redis_port || '6333')"`
 
 export PGHOST PGPORT
 
@@ -33,7 +56,13 @@ die() {
 echo "preparing postgres..."
 dropdb ${TEST_DB} # 2> /dev/null # error expected if doesn't exist, but not otherwise
 createdb -Ttemplate_postgis -EUTF8 ${TEST_DB} || die "Could not create test database"
-psql -f test.sql ${TEST_DB} 
+cat test.sql |
+  sed "s/:PUBLICUSER/${PUBLICUSER}/" |
+  sed "s/:PUBLICPASS/${PUBLICPASS}/" |
+  sed "s/:TESTUSER/${TESTUSER}/" |
+  sed "s/:TESTPASS/${TESTPASS}/" |
+  psql ${TEST_DB} 
+
 psql -f support/CDB_QueryStatements.sql ${TEST_DB} 
 psql -f support/CDB_QueryTables.sql ${TEST_DB} 
 
@@ -42,6 +71,14 @@ echo "HSET rails:users:vizzuality id 1" | redis-cli -p ${REDIS_PORT} -n 5
 echo "HSET rails:users:vizzuality database_name ${TEST_DB}" | redis-cli -p ${REDIS_PORT} -n 5
 echo "HSET rails:users:vizzuality" "map_key" "1234" | redis-cli -p ${REDIS_PORT} -n 5
 echo "SADD rails:users:vizzuality:map_key 1235" | redis-cli -p ${REDIS_PORT} -n 5
+
+# A user configured as with cartodb-2.5.0+
+echo "HSET rails:users:cartodb250user id ${TESTUSERID}" | redis-cli -p ${REDIS_PORT} -n 5
+echo 'HSET rails:users:cartodb250user database_name "'${TEST_DB}'"' | redis-cli -p ${REDIS_PORT} -n 5
+echo 'HSET rails:users:cartodb250user database_host "localhost"' | redis-cli -p ${REDIS_PORT} -n 5
+echo 'HSET rails:users:cartodb250user database_password "'${TESTPASS}'"' | redis-cli -p ${REDIS_PORT} -n 5
+echo "HSET rails:users:cartodb250user map_key 1234" | redis-cli -p ${REDIS_PORT} -n 5
+
 echo "hset rails:oauth_access_tokens:l0lPbtP68ao8NfStCiA3V3neqfM03JKhToxhUQTR consumer_key fZeNGv5iYayvItgDYHUbot1Ukb5rVyX6QAg8GaY2" | redis-cli -p ${REDIS_PORT} -n 3
 echo "hset rails:oauth_access_tokens:l0lPbtP68ao8NfStCiA3V3neqfM03JKhToxhUQTR consumer_secret IBLCvPEefxbIiGZhGlakYV4eM8AbVSwsHxwEYpzx" | redis-cli -p ${REDIS_PORT} -n 3
 echo "hset rails:oauth_access_tokens:l0lPbtP68ao8NfStCiA3V3neqfM03JKhToxhUQTR access_token_token l0lPbtP68ao8NfStCiA3V3neqfM03JKhToxhUQTR" | redis-cli -p ${REDIS_PORT} -n 3