+++ test/run_tests.sh
@@ -1,9 +1,5 @@
 #!/bin/sh
 
-# Must match redis_port in config/environments/test.js
-# TODO: read from there
-REDIS_PORT=6333
-
 # To make output dates deterministic
 export TZ='Europe/Rome'
 
@@ -13,6 +9,10 @@ OPT_DROP=yes   # drop the test environment
 cd $(dirname $0)
 BASEDIR=$(pwd)
 cd -
+
+REDIS_PORT=`node -e "console.log(require('${BASEDIR}/../config/environments/test.js').redis_port)"`
+export REDIS_PORT
+echo "REDIS_PORT: [${REDIS_PORT}]"
 
 cleanup() {
   if test x"$OPT_DROP" = xyes; then